# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PDAM Tiara Material Tracking & Verification System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/material-accountability:
    get:
      operationId: getMaterialAccountabilityReport
      tags: [reports]
      summary: Material accountability report for a period
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Accountability report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountabilityReport"
  /reports/audit-trail:
    get:
      operationId: getAuditTrail
      tags: [reports]
      summary: Audit log for a tracking or date range
      parameters:
        - name: trackingId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: actorId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit trail entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  # Upload
...
components:
  schemas:
...
    AccountabilityReport:
      type: object
      required: [period, totalIssued, totalVerified, totalKritis, totalValueIssued, totalValueVerified, rows]
      properties:
        period:
          type: object
          properties:
            from:
              type: string
              format: date
            to:
              type: string
              format: date
          required: [from, to]
        totalIssued:
          type: integer
        totalVerified:
          type: integer
        totalKritis:
          type: integer
        totalValueIssued:
          type: number
        totalValueVerified:
          type: number
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AccountabilityRow"
    AccountabilityRow:
      type: object
      required: [trackingCode, materialName, branchName, qtyIssued, qtyUsed, qtyReturned, status, issuedAt]
      properties:
        trackingCode:
          type: string
        woNumber:
          type: ["string", "null"]
        materialCode:
          type: ["string", "null"]
        materialName:
          type: string
        branchName:
          type: string
        qtyIssued:
          type: integer
        qtyUsed:
          type: ["integer", "null"]
        qtyReturned:
          type: ["integer", "null"]
        unitPrice:
          type: ["number", "null"]
        totalValue:
          type: ["number", "null"]
        status:
          type: string
        issuedAt:
          type: string
          format: date-time
        verifiedAt:
          type: ["string", "null"]
          format: date-time
    AuditLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id:
          type: string
        trackingId:
          type: ["string", "null"]
        trackingCode:
          type: ["string", "null"]
        actorId:
          type: ["string", "null"]
        actorName:
          type: ["string", "null"]
        action:
          type: string
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # Upload
...