# lib/api-zod/src/generated/types/accountabilityReport.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PDAM Tiara Material Tracking & Verification System API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountabilityReportPeriod } from './accountabilityReportPeriod';
import type { AccountabilityRow } from './accountabilityRow';

export interface AccountabilityReport {
  period: AccountabilityReportPeriod;
  totalIssued: number;
  totalVerified: number;
  totalKritis: number;
  totalValueIssued: number;
  totalValueVerified: number;
  rows: AccountabilityRow[];
}